Silence compilation warnings in semantic highlighter
Reduce warnings from 21 to 8 by prefixing unused fields/methods:
- Prefix unused llm_service, settings fields with underscore
- Prefix unused analyzer fields in AnalyticalProcessor
- Prefix unused description() methods
- Prefix unused cache field in RelationalAnalyzer
- Fix unused variable warnings with cargo fix

Remaining 8 warnings are unused variables in scaffolded LLM integration
code (future functionality) - acceptable for now.

All core functionality compiles cleanly with only minor unused variable
warnings in tier 2/3 semantic analysis (future features).

+++ src/tui/views.rs
@@ -469,7 +469,7 @@ impl IcsPanel {
             let highlighted_line = self.highlighter.highlight_line(line);
             let line_area = Rect {
                 x: inner.x,
-                y: inner.y + y_offset as u16,
+                y: inner.y + y_offset,
                 width: inner.width,
                 height: 1,
             };
